rag-engine/requirements.txt
langchain
pymupdf
pymongo
ollama
chromadb
python-dotenv
fastapi
uvicorn
fastapi
uvicorn
langchain-community
langchain-ollama